Switched the pay report to use the employee container object (#90)
* Switched the pay report to use the employee container object

* Moved old files out of the main structure and created a PayReport file

* Removed unnecessary import
